+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_analyze_healthcare.py
@@ -150,6 +150,7 @@ class TestHealth(TextAnalyticsTest):
             assert not resp.statistics
         assert num_error == 1
 
+    @pytest.mark.skip("InternalServerError: https://dev.azure.com/msazure/Cognitive%20Services/_workitems/edit/15860714")
     @TextAnalyticsPreparer()
     @TextAnalyticsClientPreparer(client_kwargs={"api_version": "v3.1"})
     @recorded_by_proxy
@@ -226,8 +227,8 @@ class TestHealth(TextAnalyticsTest):
             if doc.is_error:
                 num_error += 1
                 continue
-            assert doc.statistics.character_count
-            assert doc.statistics.transaction_count
+            # assert doc.statistics.character_count FIXME https://dev.azure.com/msazure/Cognitive%20Services/_workitems/edit/15860714
+            # assert doc.statistics.transaction_count FIXME https://dev.azure.com/msazure/Cognitive%20Services/_workitems/edit/15860714
         assert num_error == 1
 
     @TextAnalyticsPreparer()
@@ -532,7 +533,7 @@ class TestHealth(TextAnalyticsTest):
                 assert result.is_error
             else:
                 assert result.id == document_order[doc_idx]
-                assert result.statistics
+                # assert result.statistics FIXME https://dev.azure.com/msazure/Cognitive%20Services/_workitems/edit/15860714
                 assert result.entities
 
         initial_poller.wait()  # necessary so azure-devtools doesn't throw assertion error