+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -9,7 +9,9 @@
 ### Bugs Fixed
 
 ### Other Changes
+- Removed `msrest` dependency.
 - Added `typing-extensions>=4.0.1` as a dependency.
+- Added extra dependency `aio` for installing optional async dependencies. Use `pip install azure-storage-blob[aio]` to install.
 
 ## 12.14.0 (2022-10-11)